requirements.txt
# requirements.txt
# Core dependencies
google-adk==0.1.0
fastapi>=0.103.0
uvicorn>=0.23.2
pydantic>=2.3.0
litellm>=1.7.0
python-dotenv>=1.0.0
jinja2>=3.1.2

# Google AI dependencies
google-generativeai>=0.3.0
google-auth>=2.22.0

# Database for persistent sessions
# sqlite3>=3.40.0

# API integrations
httpx>=0.24.1
websockets>=11.0.3

# Data analysis and visualization
pandas>=2.0.0
matplotlib>=3.7.0
numpy>=1.24.0

# Utilities
python-multipart>=0.0.6
uuid>=1.30
pytest>=7.0.0
pytest-cov>=4.0.0
streamlit>=1.30.0
pandas>=2.0.0
plotly>=6.0.0
altair>=5.0.0
aiohttp>=3.8.0